requirements.txt
amqp==5.0.6
attrs==21.2.0
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
Click==7.0
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
flake8==3.9.2
flake8-flask==0.9.3
flake8-isort==4.1.1
Flask==1.1.4
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
greenlet==1.1.2
iniconfig==1.1.1
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
kombu==5.1.0
MarkupSafe==2.0.1
mccabe==0.6.1
packaging==21.0
pep517==0.12.0
pip-tools==6.4.0
pluggy==1.0.0
prompt-toolkit==3.0.20
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
pytest==6.2.5
python-dateutil==2.8.2
python-util==1.2.1
pytz==2021.3
r2c-py-ast==0.1.0b1
redis==3.5.3
six==1.16.0
SQLAlchemy==1.4.26
testfixtures==6.18.3
toml==0.10.2
tomli==1.2.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
WTForms==2.3.3
